Improves sql queries parsing:
Some editors append the '\n' end-of-line (eol) character at the end of a
file. In that case splitting the .sql file based on ';' separator will
lead to a group with just the eol char which is the eol just before the
eof. To avoid that confusion we can discard the last element after the
split operation and not even check for empty groups later on.