Fixes bug of missing shp2pgsql
The shp2pgsql tool is included in postgis-client package. Since this is
used in a shell pipe and is not the last command if it fails the error
is not propagated. So the check_output() is not raising an exception! To
mitigate we can check if the shp2pgsql is installed by invoking its help
menu with the shell AND operator (&&). If not present the pipe will not
be executed and an exception will be raised in the python script
indicating the problem.